+++ tests/test_board.py
@@ -101,7 +101,7 @@ class TestBoard(unittest.TestCase):
     def test_wrongTileToCell(self):
         board = Board()
         board.addTileToCell(8, 8, Tile("A", 1))
-        self.assertEqual(board.addTileToCell(8, 8, Tile(123, 1)), False)
+        self.assertIsNone(board.addTileToCell(8, 8, Tile(123, 1)))
 
 if __name__ == '__main__':
     unittest.main()